Replace direct query of consul service discovery endpoint with the official consul API

+++ core/scheduler/service_discovery.go
@@ -5,7 +5,7 @@ type Service struct {
 	Name string
 	ID   string
 	IP   string
-	Port float64
+	Port int
 	Tags []string
 }
